requirements.txt

Django==4.2.0
django-ckeditor==6.7.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-js-asset==2.1.0
django-redis==5.4.0
django-storages==1.14.2
djangorestframework==3.14.0
djoser==2.2.2
pillow==10.2.0
social-auth-app-django==5.4.0
pytz==2023.3
mutagen==1.47.0
psycopg2==2.9.9
argon2-cffi==23.1.0
stripe==7.12.0
requests==2.31.0
braintree==4.26.0
pyJWT==2.8.0
boto3==1.34.21
markdown==3.5.2
gunicorn==21.2.0
whitenoise==6.6.0